Move hooks + remaining enums into published module
