+++ libs/js/push-notifications/package.json
@@ -7,6 +7,13 @@
   "scripts": {
     "build": "tsc"
   },
+  "dependencies": {
+    "@types/react": "^18.0.28",
+    "react": "^18.2.0"
+  },
+  "devDependencies": {
+    "typescript": "^5.0.2"
+  },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/tix-factory/push-notifications.git"
@@ -20,8 +27,5 @@
   },
   "bugs": {
     "url": "https://github.com/tix-factory/push-notifications/issues"
-  },
-  "devDependencies": {
-    "typescript": "^5.0.2"
   }
 }
