+++ libs/js/push-notifications/src/enums/serviceWorkerInstallationState.ts
